nintendo/playch10.cpp: Merged playch10 source files. (#10214)
* Added finders for joystick and gun inputs.
* Fixed some bad masks (sprite DMA write and Pin-Bot).
* Made color PROM loading more concise.
* Removed pc_1942 imperfect graphics flag.  It looks as bad as it should.
* Various other minor cleanups.
